se agrego la notificacion de guardado de ubicacion

+++ lib/ubicacion.dart
@@ -3,7 +3,7 @@ import 'package:google_maps_flutter/google_maps_flutter.dart';
 class Ubicacion {
   String nombreUbicacion;
   String descripcion;
-  int calificacion;
+  double calificacion;
   double longitud, latitud;
 
   Ubicacion(
